fix: restored camera.vue
